Kill LP64-related warnings.

+++ xc/programs/Xserver/Xprint/ps/PsWindow.c
@@ -354,14 +354,14 @@ PsPaintWindow(
         }
         break;
       case GCClipMask:
-        if( (pointer)pGC->clientClipType!=(pointer)CT_NONE )
+        if( (pointer)(long)pGC->clientClipType!=(pointer)CT_NONE )
         {
           gcmask |= index;
           gcval[i++] = (pointer)CT_NONE;
         }
         break;
       case GCSubwindowMode:
-        if( (pointer)pGC->subWindowMode!=newValues[SUBWINDOW] )
+        if( (pointer)(long)pGC->subWindowMode!=newValues[SUBWINDOW] )
         {
           gcmask |= index;
           gcval[i++] = newValues[SUBWINDOW];
@@ -375,7 +375,7 @@ PsPaintWindow(
         }
         break;
       case GCFillStyle:
-        if( (pointer)pGC->fillStyle!=newValues[FILLSTYLE] )
+        if( (pointer)(long)pGC->fillStyle!=newValues[FILLSTYLE] )
         {
           gcmask |= index;
           gcval[i++] = newValues[FILLSTYLE];
